App Script
This is the code for the app. It includes a connection to an SQLite db (supposed to be for appending changes to edited dataframe but that will have to be a future iteration), a couple of dropdown widgets that will filter the data, a streamlit-aggrid table (excellent library built by Pablo Fonseca brining Javascript's AgGrid to python) which is used to display the unsummarized data behind the model (nb: the AgGrid table allows edits to be made and the edits to pass through to the training data for the model), and a plotly chart visualizing the historical demand and forecast for the next 12 months.